Oredrs Screen

+++ app/src/main/java/com/example/shipping/domain/module/Product.kt
@@ -10,6 +10,6 @@ data class Product(
     @ColumnInfo("product_name") var product_name: String,
     @ColumnInfo("product_img") var product_img: String,
     @ColumnInfo("product_description") var product_description: String,
-    @ColumnInfo("product_price") var product_price: Long,
+    @ColumnInfo("product_price") var product_price: Int,
     @ColumnInfo("product_count") var product_count: Int
 )
